Move encrypted user id and token to a module
This module is shared between filestore and submitter
because submitter will download attachments from filestore and
both needs to be the same
