+++ app/services/platform/encrypted_user_id_and_token.rb
@@ -1,0 +1,13 @@
+module Platform
+  module EncryptedUserIdAndToken
+    def encrypted_user_id_and_token
+      DataEncryption.new(
+        key: service_secret
+      ).encrypt("#{user_id}#{session[:user_token]}")
+    end
+
+    def user_id
+      session[:session_id]
+    end
+  end
+end
